requirements.txt
requests
psycopg2-binary
python-dateutil
python-dotenv
beautifulsoup4
schedule
resend
flask
gunicorn